requirements.txt
black
python-dotenv
confluent-kafka
pyyaml